% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMScoringTableReader.R
\name{AMReader}
\alias{AMReader}
\title{AMReader}
\usage{
AMReader(
  Experiment_Name = "",
  Path = getwd(),
  File_Name = "AMScorer {Experiment_Name}.xlsx",
  Condition_Include = NULL,
  Condition_Exclude = NULL,
  Structure_Include = NULL,
  Structure_Exclude = NULL,
  Facet_1 = FALSE,
  Facet_2 = FALSE,
  Facet_3 = FALSE,
  Stat_Dunn_Padj = "none",
  Stat_Wilcoxon_Padj = "none",
  Stat_Sided = "two.sided",
  Stat_Output = FALSE,
  Stat_File = "Statistical Analysis {Experiment_Name}.xlsx",
  Graph_Type = "Facets",
  Graph_Sample_Sizes = NULL,
  Graph_Condition_Order = NULL,
  Graph_Facet_1_Order = NULL,
  Graph_Facet_2_Order = NULL,
  Graph_Facet_3_Order = NULL,
  Graph_Stat_Test = NULL,
  Graph_Stat_Display = "Letters",
  Graph_Reference_Condition = NULL,
  Graph_Manual_Colour = FALSE,
  Graph_Colour = "Viridis",
  Graph_Palette = "viridis",
  Graph_Palette_Begin = 0,
  Graph_Palette_End = 1,
  Graph_Text_Colour = "grey30",
  Graph_Background_Colour = "grey95",
  Graph_Hline_Colour = "grey30",
  Graph_Legend = TRUE,
  Graph_Size_Right_Label = 1,
  Graph_Size_Top_Label = 1,
  Graph_Size_Y_Axis = 1,
  Graph_Size_X_Axis = 1,
  Graph_Size_Legend = 1,
  Graph_Size_Legend_Text = 1,
  Graph_Size_Percentages = 1,
  Graph_Size_Statistics = 1,
  Graph_Size_Datapoints = 1,
  Graph_Output = FALSE,
  Graph_Resolution = 1200,
  Graph_Width_Adjustment = 1,
  Graph_Height_Adjustment = 1,
  Graph_File = "Colonisation Percentage Graph {Experiment_Name}.png"
)
}
\arguments{
\item{Experiment_Name}{A string with the experiment name used to set the
defaults for File_Name, Stat_File, and Graph_File.}

\item{Path}{A string giving the path to the input excel file.}

\item{File_Name}{A string giving the name of the input excel file.}

\item{Condition_Include}{A vector of characters which may take values from
A to Z, AA to AZ and BA to BZ.}

\item{Condition_Exclude}{A vector of characters which may take values from
A to Z, AA to AZ and BA to BZ.}

\item{Structure_Include}{A vector of characters which may take values from
"Total", "EH", "H", "IH", "A", "V" or "S".}

\item{Structure_Exclude}{A vector of characters which may take values from
"Total", "EH", "H", "IH", "A", "V" or "S".}

\item{Facet_1}{A logical: If TRUE, data from Facet_1 will be processed.}

\item{Facet_2}{A logical: If TRUE, data from Facet_2 will be processed.}

\item{Facet_3}{A logical: If TRUE, data from Facet_3 will be processed.}

\item{Stat_Dunn_Padj}{A string which may take a value from
"none", "bonferroni", "holm", "hommel", "hochberg", "BH", "BY" or "fdr".}

\item{Stat_Wilcoxon_Padj}{A string which may take a value from
"none", "bonferroni", "holm", "hommel", "hochberg", "BH", "BY" or "fdr".}

\item{Stat_Sided}{A string which may take a value from "two.sided", "less",
or "greater".}

\item{Stat_Output}{A logical: If TRUE, the Statistical Analysis output will
be saved to excel}

\item{Stat_File}{A string giving the file name of the output from the
statistical analysis file. Does not include ".xlsx".}

\item{Graph_Type}{A string which may take a value from "Facets" or "Single".}

\item{Graph_Sample_Sizes}{A logical: If TRUE, sample size information will be
displayed on the graph.}

\item{Graph_Condition_Order}{A vector of characters which may take values
from A to Z, AA to AZ and BA to BZ.}

\item{Graph_Facet_1_Order}{A vector of characters giving the possible values
of Facet_1 in the "Conditions" tab of the associated excel input file.}

\item{Graph_Facet_2_Order}{A vector of characters giving the possible values
of Facet_2 in the "Conditions" tab of the associated excel input file.}

\item{Graph_Facet_3_Order}{A vector of characters giving the possible values
of Facet_3 in the "Conditions" tab of the associated excel input file.}

\item{Graph_Stat_Test}{A string which may take a value from
"Wilcoxon", "Tukey", "Dunn", or "NULL".}

\item{Graph_Stat_Display}{A string which may take a value from
"Letters" or "Reference"}

\item{Graph_Reference_Condition}{A string which may take a values from
"Letters" or "Reference"}

\item{Graph_Manual_Colour}{A logical: If TRUE, the input from Manual_Colour
given in the "Conditions" tab of the associated excel input file.}

\item{Graph_Colour}{A string which may take a values from
"Viridis", "Brewer", "Grey"}

\item{Graph_Palette}{A string which may take any value which defines a valid
palette from Viridis or Brewer colour schemes. See the R documentation
associated with both colour schemes.}

\item{Graph_Palette_Begin}{A numeric taking values between 0 and 1.}

\item{Graph_Palette_End}{A numeric taking values between 0 and 1.}

\item{Graph_Text_Colour}{A string which may take any valid colour input
listed in color() or any hexcode.}

\item{Graph_Background_Colour}{A string which may take any valid colour input
listed in color() or any hexcode.}

\item{Graph_Hline_Colour}{A string which may take any valid colour input
listed in color() or any hexcode.}

\item{Graph_Legend}{A logical: If FALSE, the graph legend will not be
displayed}

\item{Graph_Size_Right_Label}{A numeric value greater than 0. Multiplies
the size of the "Structure" labels.}

\item{Graph_Size_Top_Label}{A numeric value greater than 0. Multiplies
the size of the "Facet" labels.}

\item{Graph_Size_Y_Axis}{A numeric value greater than 0. Multiplies
the size of the "Y_axis" numbers}

\item{Graph_Size_X_Axis}{A numeric value greater than 0. Multiplies
the size of the "X_axis" text.}

\item{Graph_Size_Legend}{A numeric value greater than 0. Multiplies
the size of the figure legend.}

\item{Graph_Size_Legend_Text}{A numeric value greater than 0. Multiplies
the size of the "Legend" text.}

\item{Graph_Size_Percentages}{A numeric value greater than 0. Multiplies
the size of the Y-label "Percentage Colonisation" text.}

\item{Graph_Size_Statistics}{A numeric value greater than 0. Multiplies
the size of the "statistical information" text.}

\item{Graph_Size_Datapoints}{A numeric value greater than 0. Multiplies
the size of the "data points".}

\item{Graph_Output}{A logical: If TRUE, the graph will be saved as an
output}

\item{Graph_Resolution}{A string giving the desired dpi of the output
graph file}

\item{Graph_Width_Adjustment}{A string giving the desired numeric
for stretching the graph output's width. Use only if necessary.}

\item{Graph_Height_Adjustment}{A string giving the desired numeric
for stretching the graph output's height. Use only if necessary.}

\item{Graph_File}{A string giving the file name of the graph output.
Specify the file extension (.png, .pdf, .jpeg etc.)}
}
\description{
AMReader takes data collected in AMScorer, processes
it, performs statistical analyses and enables production and customisation of
a graphical output. Please see README for further information.
}
\details{
See the README for further information.
}
